Extracts derived quantities from SS output

A number of derived quantities are available in the Report.sso file, and are useful for checking that the generated FLR objects lead to the same values. These functions are called on the list returned by readOutputss3 or r4ss::SS_output, and extract the yearly values from the following rows in the derived_quants data.frame:

extractSSB(out, endyr = sum(c(out$endyr, out$nforecastyears), na.rm = TRUE))

extractRec(out, endyr = sum(c(out$endyr, out$nforecastyears), na.rm = TRUE))

extractFbar(out, endyr = sum(c(out$endyr, out$nforecastyears), na.rm = TRUE))

extractZatage(out)

extractDevs(out)

Arguments

out

A list as returned by r4ss::SS_output.

Value

An FLQuant object of the requested quantity.

Details

extractSSB: SSB_y for y between startyr and endyr.

extractRec: Recr_y for y between startyr and endyr.

extractFbar: F_y for y between startyr and endyr.

extractZatage: Z_ay for y between startyr and endyr and ages but last.

For 2 sex models (nsexes), extractRec will return a two-unit FLQuant, with the Recr_y values split according to the recruitment sex ratio. This is extracted from recruitment_dist[[1]][, "Frac/sex"].

The value returned by extractFbar is the actual mean F over the age range, and the value in derived_quants is corrected according to F_std_basis for models where F reporting basis is F/FMSY.

Examples

out <- readOutputss3(system.file("ext-data", "herring", package="ss3om"))

extractSSB(out)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1963    1964    1965    1966    1967    1968    1969    1970    1971   
#>   all 1115500 1113640 1088450  999404  913065  824116  729948  670133  513631
#>      year
#> age   1972    1973    1974    1975    1976    1977    1978    1979    1980   
#>   all  545944  584478  500793  530701  536931  575989  649982  594032  520278
#>      year
#> age   1981    1982    1983    1984    1985    1986    1987    1988    1989   
#>   all  530449  548494  598526  655894  719889  808423  895734  869834  995561
#>      year
#> age   1990    1991    1992    1993    1994    1995    1996    1997    1998   
#>   all 1124340 1255360 1220700 1175460 1298200 1146920 1131740  947283  903654
#>      year
#> age   1999    2000    2001    2002    2003    2004    2005    2006    2007   
#>   all  881440  819745  804034  812500  824125  864074  876191  783562  763210
#>      year
#> age   2008    2009    2010    2011    2012    2013    2014    2015    2016   
#>   all  726905  723815  879936  787291  804124  819487  768625  739417  702689
#>      year
#> age   2017    2018   
#>   all  707672  779999
#> 
#> units:  t 

extractFbar(out)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1964   1965   1966   1967   1968   1969   1970   1971   1972   1973  
#>   all 0.0244 0.0278 0.0241 0.0329 0.0380 0.0537 0.0534 0.0776 0.0668 0.0484
#>      year
#> age   1974   1975   1976   1977   1978   1979   1980   1981   1982   1983  
#>   all 0.0550 0.0424 0.0588 0.0594 0.0606 0.0488 0.0608 0.0415 0.0521 0.0443
#>      year
#> age   1984   1985   1986   1987   1988   1989   1990   1991   1992   1993  
#>   all 0.0510 0.0486 0.0473 0.0399 0.0421 0.0349 0.0391 0.0308 0.0432 0.0452
#>      year
#> age   1994   1995   1996   1997   1998   1999   2000   2001   2002   2003  
#>   all 0.0557 0.0627 0.0630 0.0798 0.0736 0.0819 0.0847 0.0837 0.0760 0.0706
#>      year
#> age   2004   2005   2006   2007   2008   2009   2010   2011   2012   2013  
#>   all 0.0746 0.0804 0.1003 0.1148 0.1043 0.1085 0.1069 0.1228 0.1632 0.1835
#>      year
#> age   2014   2015   2016   2017   2018  
#>   all 0.1953 0.2073 0.2422 0.1965 0.1787
#> 
#> units:  f