A function to read SS3 results as an FLStock object

Results of a run of the Stock Synthesis sofware, SS3 (Methot & Wetzel, 2013), can be loaded into an object of class FLStock. The code makes use of the r4ss::SS_output function to obtain a list from Report.sso. The following elements of that list are used to generate the necessary information for the slots in FLStock: "catage", "natage", "ageselex", "endgrowth", "catch_units", "nsexes", "nseasons", "nareas", "IsFishFleet", "fleet_ID", "FleetNames", "spawnseas", "inputs" and "SS_version".

readFLSss3(dir, birthseas = out$spawnseas, name = "",
  desc = paste(out$inputs$repfile, out$SS_version, sep = " - "),
  repfile = "Report.sso", compfile = "CompReport.sso")

Arguments

dir

Directory holding the SS3 output files

birthseas

Birth seasons for this stock, defaults to spawnseas

name

Name of the output object to fil the name slot

desc

Description of the output object to fill the desc slot

Value

An object of class FLStock

References

Methot RD Jr, Wetzel CR (2013) Stock Synthesis: A biological and statistical framework for fish stock assessment and fishery management. Fisheries Research 142: 86-99.

See also

FLComp

Examples