Functions to convert SS3 output into FLQuant(s)

A series of auxiliary functions that convert one or more elements, typically of class data.frame. in the list returned by r4ss::SS_output into particular FLQuant or FLQuants objects.

ss3index(cpue, fleets)

ss3index.res(cpue, fleets)

ss3index.var(cpue, fleets)

ss3index.q(cpue, fleets)

ss3sel.pattern(selex, years, fleets, morphs, factor = "Asel2")

ss3wt(endgrowth, dmns, birthseas)

ss3mat(endgrowth, dmns, birthseas, option = 3)

ss3m(endgrowth, dmns, birthseas)

ss3n(n, dmns, birthseas)

ss3catch(catage, wtatage, dmns, birthseas, idx)

Arguments

cpue
A data frame obtained from SS_output$cpue.

fleets
A named list of vector of the fleets to be extracted.

selex
A data frame obtained from SS_output$ageselex.

years
Vector of years for which the index applies

morphs
Vector of morphs to use

endgrowth
A data frame obtained from SS_output$endgrowth.

dmns
dimnames of the output object, usually obatined using getDimnames.

birthseas
The birthseasons for this stock as a numeric vector.

n
A data frame obtained from SS_output$natage.

catage
A data frame obtained from SS_output$catage.

wtatage
A data frame obtained from SS_output$endgrowth but subset for birthseas and RetWt:_idx.

idx
The fishing fleets, as in SS_output$fleet_ID[SS_output$IsFishFleet].

fleets
Named vector of fleets (numeric) codes

Value

An FLQuant or FLQuants object, depending on the converted data structure

Details

ss3index returns the index slot of each survey/CPUE fleet.

ss3index.res returns the index.res slot of each survey/CPUE fleet.

ss3index.var returns the index.var slot of each survey/CPUE fleet.

ss3index.q returns the index.q slot of each survey/CPUE fleet.

ss3sel.pattern returns the sel.pattern slot of each survey/CPUE fleet.

ss3wt returns the stock.wt slot.

ss3mat returns the mat slot.

ss3m returns the m slot.

ss3m returns the m slot.

ss3catch currently returns the landings.n slot, equal to catch.n as discards are not being parsed.

See also