Extracts derived quantities from SS output

A number of derived quantities are available in the Report.sso file, and are useful for checking that the generated FLR objects lead to the same values. These functions are called on the list returned by readOutputss3 or r4ss::SS_output, and extract the yearly values from the following rows in the derived_quants data.frame:

extractSSB(out)

extractRec(out)

extractFbar(out)

extractZatage(out)

extractDevs(out)

Arguments

out
A list as returned by r4ss::SS_output.

Value

An FLQuant object of the requested quantity.

Details

extractSSB: SSB_y for y between startyr and endyr.

extractRec: Recr_y for y between startyr and endyr.

extractFbar: F_y for y between startyr and endyr.

extractZatage: Z_ay for y between startyr and endyr and ages but last.

For 2 sex models (nsexes), extractRec will return a two-unit FLQuant, with the Recr_y values split according to the recruitment sex ratio. This is extracted from recruitment_dist[[1]][, "Frac/sex"].

The value returned by extractFbar is the actual mean F over the age range, and the value in derived_quants is corrected according to F_report_basis for models where F reporting basis is F/FMSY.

Examples

out <- readOutputss3(system.file("ext-data", "herring", package="ss3om"))

extractSSB(out)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1963    1964    1965    1966    1967    1968    1969    1970    1971   
#>   all 1115500 1113640 1088450  999404  913065  824116  729948  670133  513631
#>      year
#> age   1972    1973    1974    1975    1976    1977    1978    1979    1980   
#>   all  545944  584478  500793  530701  536931  575989  649982  594032  520278
#>      year
#> age   1981    1982    1983    1984    1985    1986    1987    1988    1989   
#>   all  530449  548494  598526  655894  719889  808423  895734  869834  995561
#>      year
#> age   1990    1991    1992    1993    1994    1995    1996    1997    1998   
#>   all 1124340 1255360 1220700 1175460 1298200 1146920 1131740  947283  903654
#>      year
#> age   1999    2000    2001    2002    2003    2004    2005    2006    2007   
#>   all  881440  819745  804034  812500  824125  864074  876191  783562  763210
#>      year
#> age   2008    2009    2010    2011    2012    2013    2014    2015    2016   
#>   all  726905  723815  879936  787291  804124  819487  768625  739417  702689
#>      year
#> age   2017    2018   
#>   all  707672  779999
#> 
#> units:  t 

extractFbar(out)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1964     1965     1966     1967     1968     1969     1970     1971    
#>   all 0.024390 0.027801 0.024055 0.032882 0.038037 0.053698 0.053383 0.077643
#>      year
#> age   1972     1973     1974     1975     1976     1977     1978     1979    
#>   all 0.066840 0.048423 0.055038 0.042395 0.058760 0.059356 0.060590 0.048773
#>      year
#> age   1980     1981     1982     1983     1984     1985     1986     1987    
#>   all 0.060842 0.041520 0.052141 0.044258 0.051032 0.048641 0.047333 0.039914
#>      year
#> age   1988     1989     1990     1991     1992     1993     1994     1995    
#>   all 0.042073 0.034875 0.039074 0.030751 0.043203 0.045203 0.055656 0.062677
#>      year
#> age   1996     1997     1998     1999     2000     2001     2002     2003    
#>   all 0.062953 0.079769 0.073551 0.081917 0.084717 0.083675 0.076044 0.070579
#>      year
#> age   2004     2005     2006     2007     2008     2009     2010     2011    
#>   all 0.074583 0.080404 0.100346 0.114826 0.104348 0.108547 0.106926 0.122755
#>      year
#> age   2012     2013     2014     2015     2016     2017     2018    
#>   all 0.163153 0.183518 0.195297 0.207305 0.242240 0.196540 0.178681
#> 
#> units:  f